requirements.txt
tensorflow
streamlit
numpy
<<<<<<< HEAD
Pillow
=======
Pillow
>>>>>>> 4705dae252fc4505977c8c32b9e75d13adec5e4f
